+++ src/codemagic/apple/app_store_connect/builds/builds.py
@@ -1,6 +1,7 @@
 from dataclasses import dataclass
 from typing import List
 from typing import Optional
+from typing import Sequence
 from typing import Tuple
 from typing import Type
 from typing import TypeVar
@@ -9,12 +10,14 @@ from typing import Union
 from codemagic.apple.app_store_connect.resource_manager import ResourceManager
 from codemagic.apple.resources import App
 from codemagic.apple.resources import AppStoreVersion
+from codemagic.apple.resources import BetaReviewState
 from codemagic.apple.resources import Build
 from codemagic.apple.resources import BuildProcessingState
 from codemagic.apple.resources import LinkedResourceData
 from codemagic.apple.resources import PreReleaseVersion
 from codemagic.apple.resources import Resource
 from codemagic.apple.resources import ResourceId
+from codemagic.apple.resources import ResourceType
 
 IncludedResource = TypeVar('IncludedResource', bound=Resource)
 
@@ -35,6 +38,7 @@ class Builds(ResourceManager[Build]):
         expired: Optional[bool] = None
         id: Optional[ResourceId] = None
         processing_state: Optional[BuildProcessingState] = None
+        beta_app_review_submission_beta_review_state: Optional[Union[BetaReviewState, Sequence[BetaReviewState]]] = None
         version: Optional[Union[str, int]] = None
         pre_release_version_version: Optional[str] = None
 
@@ -42,12 +46,15 @@ class Builds(ResourceManager[Build]):
         def _get_field_name(cls, field_name) -> str:
             if field_name == 'pre_release_version_version':
                 field_name = 'pre_release_version.version'
+            elif field_name == 'beta_app_review_submission_beta_review_state':
+                field_name = 'beta_app_review_submission.beta_review_state'
             return super()._get_field_name(field_name)
 
     class Ordering(ResourceManager.Ordering):
         PRE_RELEASE_VERSION = 'preReleaseVersion'
         UPLOADED_DATE = 'uploadedDate'
         VERSION = 'version'
+        BETA_REVIEW_STATE = 'betaReviewState'
 
     def read(self, build: Union[LinkedResourceData, ResourceId]) -> Build:
         """
@@ -57,10 +64,12 @@ class Builds(ResourceManager[Build]):
         response = self.client.session.get(f'{self.client.API_URL}/builds/{build_id}').json()
         return Build(response['data'])
 
-    def list(self,
-             resource_filter: Filter = Filter(),
-             ordering=Ordering.UPLOADED_DATE,
-             reverse=False) -> List[Build]:
+    def list(
+        self,
+        resource_filter: Filter = Filter(),
+        ordering=Ordering.UPLOADED_DATE,
+        reverse=False,
+    ) -> List[Build]:
         """
         https://developer.apple.com/documentation/appstoreconnectapi/list_builds
         """
@@ -126,6 +135,31 @@ class Builds(ResourceManager[Build]):
         ).json()
 
         return Build(response['data']), include_type(response['included'][0])
+
+    def modify(
+        self,
+        build: Union[LinkedResourceData, ResourceId],
+        expired: Optional[bool] = None,
+    ) -> Build:
+        """
+        https://developer.apple.com/documentation/appstoreconnectapi/modify_a_build
+        """
+        build_id = self._get_resource_id(build)
+
+        attributes = {}
+        if expired is not None:
+            attributes['expired'] = expired
+
+        payload = self._get_update_payload(
+            build_id,
+            ResourceType.BUILDS,
+            attributes=attributes,
+        )
+        response = self.client.session.patch(
+            f'{self.client.API_URL}/builds/{build_id}',
+            json=payload,
+        ).json()
+        return Build(response['data'])
 
     @classmethod
     def _get_include_field_name(cls, include_type: Type[IncludedResource]) -> str: