+++ src/codemagic/apple/app_store_connect/versioning/review_submissions.py
@@ -1,4 +1,7 @@
+from dataclasses import dataclass
+from typing import List
 from typing import Optional
+from typing import Sequence
 from typing import Type
 from typing import Union
 
@@ -9,6 +12,7 @@ from codemagic.apple.resources import Platform
 from codemagic.apple.resources import ResourceId
 from codemagic.apple.resources import ResourceType
 from codemagic.apple.resources import ReviewSubmission
+from codemagic.apple.resources import ReviewSubmissionState
 
 
 class ReviewSubmissions(ResourceManager[ReviewSubmission]):
@@ -20,6 +24,12 @@ class ReviewSubmissions(ResourceManager[ReviewSubmission]):
     @property
     def resource_type(self) -> Type[ReviewSubmission]:
         return ReviewSubmission
+
+    @dataclass
+    class Filter(ResourceManager.Filter):
+        app: Optional[Union[ResourceId, str]] = None
+        platform: Optional[Platform] = None
+        state: Optional[Union[ReviewSubmissionState, Sequence[ReviewSubmissionState]]] = None
 
     def create(
         self,
@@ -54,6 +64,16 @@ class ReviewSubmissions(ResourceManager[ReviewSubmission]):
         review_submission_id = self._get_resource_id(review_submission)
         response = self.client.session.get(f'{self.client.API_URL}/reviewSubmissions/{review_submission_id}').json()
         return ReviewSubmission(response['data'])
+
+    def list(self, resource_filter: Filter = Filter()) -> List[ReviewSubmission]:
+        """
+        https://developer.apple.com/documentation/appstoreconnectapi/get_v1_reviewsubmissions
+        """
+        review_submissions = self.client.paginate(
+            f'{self.client.API_URL}/reviewSubmissions',
+            params=resource_filter.as_query_params(),
+        )
+        return [ReviewSubmission(submission_info) for submission_info in review_submissions]
 
     def modify(
         self,