+++ src/codemagic/tools/_app_store_connect/action_groups/review_submissions_actions_group.py
@@ -46,7 +46,7 @@ class ReviewSubmissionsActionGroup(AbstractBaseAction, metaclass=ABCMeta):
         self,
         review_submission_id: ResourceId,
         should_print: bool = True,
-    ):
+    ) -> ReviewSubmission:
         """
         Read Review Submission information
         """
@@ -67,7 +67,7 @@ class ReviewSubmissionsActionGroup(AbstractBaseAction, metaclass=ABCMeta):
         should_print: bool = True,
     ) -> ReviewSubmission:
         """
-        Discard review submission from App Review
+        Discard a specific review submission from App Review
         """
         return self._modify_resource(
             self.api_client.review_submissions,